Fix astronomical price calculation bug
- Change backend query to cast median to integer to prevent numeric overflow
- Replace frontend average calculation with median display to avoid compounding errors
- Frontend now shows 'Median Price' instead of calculated average
- This should resolve the  quintillion price display issue

+++ server/index.ts
@@ -683,7 +683,7 @@ app.get('/api/insights/price-trends', requireAuth, async (req, res) => {
       WITH recent_data AS (
         SELECT
           TO_CHAR(active_month, 'YYYY-MM') as month,
-          ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price_search_sold)::numeric, 0) as avg_price,
+          ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price_search_sold), 0)::integer as avg_price,
           COUNT(*) as total_sales,
           active_month
         FROM property_sales
